Rewrite Resources page
The prior implementation required a mess of shortcodes that couldn't be
nested, thus obscuring the structure of the collapsible menus. Using
shortcode calls with the percent symbols converts inner markdown to
HTML,  which can be processed via Regex in the absence of a better way
of traversing HTML structures within shortcodes. The list HTML tags
are replaced with tags for collapsible elements, then this is piped to
a partial that wraps each collapsible menu in a pureCSS grid.